requirements.txt
ipdb
ipykernel
matplotlib
numpy
pandas
seaborn
scikit-learn
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
google-cloud-bigquery<3.0.0
google-cloud-storage
pyarrow
pywt
scipy


python-dotenv
fastapi
pytz
uvicorn

wfdb
librosa

mlflow
prefect
